Factoring bloom filter types out of the protocol folder
