+++ Network/Haskoin/Protocol/Message.hs
@@ -31,6 +31,7 @@ import Network.Haskoin.Transaction.Types
 import Network.Haskoin.Block.Types
 import Network.Haskoin.Crypto.Hash 
 import Network.Haskoin.Crypto.Merkle
+import Network.Haskoin.Crypto.Bloom
 import Network.Haskoin.Constants
 import Network.Haskoin.Util
