+++ tests/Network/Haskoin/Crypto/Bloom/Tests.hs
@@ -3,20 +3,24 @@ module Network.Haskoin.Crypto.Bloom.Tests (tests) where
 import Test.Framework (Test, testGroup)
 import Test.Framework.Providers.QuickCheck2 (testProperty)
 
-import Network.Haskoin.Protocol.Arbitrary ()
+import Data.Binary
+import Data.Binary.Get
+import Data.Binary.Put
+
+import Network.Haskoin.Crypto.Arbitrary ()
 import Network.Haskoin.Crypto
 import Network.Haskoin.Util
 
 tests :: [Test]
 tests = 
-    [ testGroup "Bloom Filters"
-        [ testProperty "decode . encode BloomFilter" decodeEncodeBloom
+    [ testGroup "Serialize & de-serialize protocol messages"
+        [ testProperty "BloomFlags" (metaBinary :: BloomFlags -> Bool)
+        , testProperty "BloomFilter" (metaBinary :: BloomFilter -> Bool)
+        , testProperty "FilterLoad" (metaBinary :: FilterLoad -> Bool)
+        , testProperty "FilterAdd" (metaBinary :: FilterAdd -> Bool)
         ]
     ]
 
-{- Bloom Filters -}
-
-decodeEncodeBloom :: BloomFilter -> Bool
-decodeEncodeBloom bfilter = (decode' $ encode' bfilter) == bfilter
-
+metaBinary :: (Binary a, Eq a) => a -> Bool
+metaBinary x = (decode' $ encode' x) == x
